requirements.txt
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.5
Pillow==5.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
requests==2.20.0
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
sklearn==0.20.0
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
pyyaml==3.13
